backend/requirements.txt
# Core dependencies for RAG application
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
pypdf==3.17.4

# Web framework
fastapi==0.104.1
python-multipart==0.0.6
